FixedSomeBugs
Hydro_veg now basically works, still giving concentrations that are off but that may be more due to denominator (root mass/volume) than to numerator. Or possibly I need to re-look at the equations for the Z values, not sure what the best solution might be if not a denominator problem. Should check the mass balance from their work - can  see if my total mass matches what they have or if I can calculate the volumes from that!